1.优化gradle文件，引入config.gradle统一配置管理 2.把滤镜库抽出来成一个module，并引入Instagram的滤镜 3.拦截器加入版本号，并清除加密算法

+++ app/src/main/java/com/miguan/yjy/model/services/Services.java
@@ -345,7 +345,7 @@ public interface Services {
      */
     @GET("?action=queryCosmetics")
     Observable<UserProduct> queryCode(
-            @Query("id") int brandId,
+            @Query("id") Long brandId,
             @Query("number") String number
     );
 
@@ -355,7 +355,7 @@ public interface Services {
     @GET("?action=addRemind")
     Observable<String> addRepository(
             @Query("user_id") int userId,
-            @Query("brand_id") int brandId,
+            @Query("brand_id") Long brandId,
             @Query("brand_name") String brand_name,
             @Query("product") String product,
             @Query("is_seal") int is_seal,
